# Optimize for sequential abbreviation codes