--- a/src/unit.rs
+++ b/src/unit.rs
@@ -2924,7 +2924,9 @@ mod tests {
                    Ok((EndianBuf::new(expected_rest), expected_unit)));
     }
 
-    fn section_contents<F>(f: F) -> Vec<u8> where F: Fn(Section) -> Section {
+    fn section_contents<F>(f: F) -> Vec<u8>
+        where F: Fn(Section) -> Section
+    {
         f(Section::with_endian(Endian::Little)).get_contents().unwrap()
     }
 
@@ -3012,8 +3014,7 @@ mod tests {
                      (AttributeValue::Data(p1i64), Some(1), Some(1)),
                      (AttributeValue::Data(n1i64), Some(std::u64::MAX), Some(-1)),
                      (AttributeValue::Sdata(1), None, Some(1)),
-                     (AttributeValue::Udata(1), Some(1), None),
-        ];
+                     (AttributeValue::Udata(1), Some(1), None)];
         for test in tests.iter() {
             let (value, expect_udata, expect_sdata) = *test;
             let attribute = Attribute {