Normally abbreviations will have sequential codes starting at 1. Optimize for this by storing them in a Vec.

    Before:
    test bench_parsing_debug_abbrev                ... bench:      19,021 ns/iter (+/- 794)
    test bench_parsing_debug_info                  ... bench:   2,737,211 ns/iter (+/- 92,580)

    After:
    test bench_parsing_debug_abbrev                ... bench:      14,844 ns/iter (+/- 628)
    test bench_parsing_debug_info                  ... bench:   2,326,573 ns/iter (+/- 29,075)

This is a bit of a trade-off if the codes aren't sequential, but I think the cost is small relative to the cost of hashing, and the potential gain is large.