@philipc 2016-10-07
I probably should add some tests for all the edge cases in this too...